Create i need help
First of all, hello, I wish you good work.
If the story mode of the game is played, there is an explosion scene during the game, the system is activated in that scene, then we are given something like a car and we take it to the area where the productive experiment will be carried out. They are asked to push. However, it does not work, the vehicle cannot be pushed in any way.
I deleted the game and reinstalled it, it didn't work. I completely uninstalled the game and tried again, but it still doesn't work. There is a solution system on the internet for this, but it still does not appear. I request urgent help regarding this issue.